refactor SMTC 2024 passenger campaign
