+++ api/src/pdc/services/policy/engine/policies/20240101_SMTC2024Passenger.ts
@@ -12,51 +12,43 @@ import {
   isOperatorOrThrow,
   onDistanceRange,
   onDistanceRangeOrThrow,
+  perKm,
   perSeat,
   watchForGlobalMaxAmount,
-  watchForPersonMaxAmountByMonth,
 } from '../helpers';
-import { TimestampedOperators, getOperatorsAt } from '../helpers/getOperatorsAt';
-import { watchForPersonMaxAmountByYear, watchForPersonMaxTripByDay } from '../helpers/limits';
-import { description } from './20240101_SMTC.html';
+import { watchForPersonMaxTripByDay } from '../helpers/limits';
 import { AbstractPolicyHandler } from './AbstractPolicyHandler';
+import { description } from './20240101_SMTC2024Passenger.html';
 
 // Politique Syndicat Mixte des Transports en Commun de l’Agglomération Clermontoise (SMTC)
 // aom = 256300120
-export const SMTC2024: PolicyHandlerStaticInterface = class
+export const SMTC2024Passenger: PolicyHandlerStaticInterface = class
   extends AbstractPolicyHandler
   implements PolicyHandlerInterface
 {
-  static readonly id = 'smtc_2024';
-  protected operators: TimestampedOperators = [
-    {
-      date: new Date('2024-01-01T00:00:00+0100'),
-      operators: [OperatorsEnum.MOV_ICI],
-    },
-  ];
+  static readonly id = 'smtc_2024_passenger';
+  protected operators = [OperatorsEnum.MOV_ICI];
   protected operator_class = ['B', 'C'];
 
   constructor(public max_amount: number) {
     super();
     this.limits = [
-      ['AFE1C47D-BF05-4FA9-9133-853D29797D09', 90_00, watchForPersonMaxAmountByMonth, LimitTargetEnum.Driver],
-      ['AFE1C47D-BF05-4FA9-9133-853D2987GF56', 540_00, watchForPersonMaxAmountByYear, LimitTargetEnum.Driver],
-      ['AFE1C47D-BF05-4FA9-9133-853D297AZEPD', 4, watchForPersonMaxTripByDay, LimitTargetEnum.Driver],
-      ['98B26189-C6FC-4DB1-AC1C-41F779C5B3C7', this.max_amount, watchForGlobalMaxAmount],
+      ['be9c45cf-5e07-4c37-b8bf-59601e3607f7', 2, watchForPersonMaxTripByDay, LimitTargetEnum.Passenger],
+      ['4a17073a-e8d7-40ba-bcc2-26d126b41ca4', this.max_amount, watchForGlobalMaxAmount],
     ];
   }
 
   protected slices: RunnableSlices = [
     {
       start: 5_000,
-      end: 80_000,
-      fn: (ctx: StatelessContextInterface) => perSeat(ctx, 150),
+      end: 30_001,
+      fn: (ctx: StatelessContextInterface) => perSeat(ctx, perKm(ctx, { amount: 10 })),
     },
   ];
 
   protected processExclusion(ctx: StatelessContextInterface) {
-    isOperatorOrThrow(ctx, getOperatorsAt(this.operators, ctx.carpool.datetime));
-    onDistanceRangeOrThrow(ctx, { min: 5_000, max: 80_000 });
+    isOperatorOrThrow(ctx, this.operators);
+    onDistanceRangeOrThrow(ctx, { min: 5_000, max: 30_001 });
     isOperatorClassOrThrow(ctx, this.operator_class);
   }
 
@@ -79,7 +71,7 @@ export const SMTC2024: PolicyHandlerStaticInterface = class
     return {
       tz: 'Europe/Paris',
       slices: this.slices,
-      operators: getOperatorsAt(this.operators),
+      operators: this.operators,
       limits: {
         glob: this.max_amount,
       },
