+++ api/src/pdc/services/policy/engine/policies/20240101_SMTCDriver.spec.ts
@@ -2,7 +2,7 @@ import test from 'ava';
 import { v4 } from 'uuid';
 import { OperatorsEnum } from '../../interfaces';
 import { makeProcessHelper } from '../tests/macro';
-import { SMTC2024 as Handler } from './20240101_SMTC';
+import { SMTC2024Driver as Handler } from './20240101_SMTCDriver';
 
 const defaultPosition = {
   arr: '74278',
